+++ serverMiddleware/api.js
@@ -1,10 +1,15 @@
 const dbAdapter = require('./rethinkdb')
 const express = require('express')
-const axios = require('axios')
-const moment = require('moment')
 const bodyParser = require('body-parser')
-const config = require('../config')
 const app = express()
+const { validateToken } = require('./middleware')
+
+const {
+  formatGamesFromUrl,
+  getOutcome,
+  dataCollectedToday,
+  getDataFromSportsRadar
+} = require('./apiUtils')
 
 app.use(bodyParser.json())
 
@@ -25,6 +30,8 @@ app.get(`/bets`, validateToken, (req, res) => {
 
 app.get(`/games`, async (req, res) => {
   if (await dataCollectedToday()) {
+    console.log('lolg')
+    
     return dbAdapter.getGamesByWeek()
       .then(games => games.reduce(formatGamesFromUrl, {}))
       .then(games => {
@@ -40,69 +47,5 @@ app.get(`/games`, async (req, res) => {
       })
     })
 })
-
-function dataCollectedToday () {
-  // const today = moment().format('YYYY-MM-DD')
-  return true
-  // return dbAdapter.getCollectedDate()
-  //   .then(date => today === date)
-}
-
-function getDataFromSportsRadar () {
-  console.log('GETTN DATA!')
-  return axios.get(`https://api.sportradar.us/nfl-ot2/games/2017/reg/schedule.json?api_key=${config.sportRadar.api_key}`)
-    .then(res => res.data.weeks)
-    .then(weeks => weeks.forEach(week => {
-      const games = week.games.map(game =>
-        Object.assign(game, { week: moment(game.scheduled).week() })
-      )
-      dbAdapter.insertGames(games)
-    }))
-    // .then(updateCollectedTimestamp)
-    .then(dbAdapter.getGamesByWeek)
-    .then(games => games.reduce(formatGamesFromUrl, {}))
-    .catch(console.log)
-}
-
-function getOutcome (game) {
-  let outcome = game.scoring
-  if (outcome) {
-    let diff = outcome.home_points - outcome.away_points
-    if (diff > 0) {
-      return 'home'
-    } else if (diff < 0) {
-      return 'away'
-    } else {
-      return 'tie'
-    }
-  }
-}
-
-function formatGamesFromUrl (formattedGames, gameGroup, weekNo) {
-  formattedGames[gameGroup.group] = gameGroup.reduction.map(game => ({
-    id: game.id,
-    status: game.status,
-    playDate: game.scheduled,
-    homeTeam: game.home.name,
-    awayTeam: game.away.name,
-    deadlineDate: game.scheduled,
-    outcome: getOutcome(game)
-  }))
-  return formattedGames
-}
-
-async function validateToken (req, res, next) {
-  const token = req.headers.authorization
-  let emailFromToken
-  delete req.username
-  try {
-    const result = await axios.get(`https://www.googleapis.com/oauth2/v3/tokeninfo?id_token=${token}`)
-    emailFromToken = result.data.email
-  } catch (e) {
-    throw new Error(`Couldnt verify token ${e}`)
-  }
-  req.username = emailFromToken
-  next()
-}
 
 module.exports = app
