+++ utils.js
@@ -2,28 +2,21 @@
 const axios = require('axios')
 const config = require('./config')
 
-function updateCollectedTimestamp () {
-  // const today = moment().format('YYYY-MM-DD')
-  // return dbAdapter.setCollectedDate(today)
-}
-
 async function getGames (token) {
-  let games
-  games = await axios.get(`http://${config.serverUrl}/games`)
-    .then(res => res.data.games)
-    // const username = await dbAdapter.getUserFromAuth(token)
-    // const userBets = await dbAdapter.getUserBets(username)
-  console.log('sss', JSON.stringify(games))
+  console.log('lol')
 
+  const [games] = await Promise.all([
+    axios.get(`http://${config.serverUrl}/games`).then(res => res.data.games)
+    // axios.get(`http://${config.serverUrl}/bets`)
+  ])
   return games
   // return games.map(game => {
   //   Object.assign({}, game,
-  //     userBets.find(bet => bet.id === game.id)
+  //     bets.find(bet => bet.id === game.id)
   //   )
   // })
 }
 
 export {
-  getGames,
-  updateCollectedTimestamp
+  getGames
 }
